FileList作成処理をListViewクラスから分離
