+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/EditorScript/FileList/FileListData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f833fa7f9aa504541a731bde52094bfb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
