+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/EditorScript/FileList/FileListNode.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using UnityEngine;
 using System.IO;
 #if UNITY_EDITOR
 using UnityEditor;
