+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/Utils.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6f09306d2153c5a4dad15ae6d3e2cab3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
